refactor(dompurify): unify sanitisation usages
