+++ dotcom-rendering/index.d.ts
@@ -1292,10 +1292,6 @@ type AdSlotType =
 // 3rd party type declarations //
 // ------------------------------
 /* eslint-disable @typescript-eslint/no-explicit-any */
-declare module 'dompurify' {
-	const createDOMPurify: any;
-	export default createDOMPurify;
-}
 declare module 'compose-function' {
 	const compose: any;
 	export default compose;
