+++ dotcom-rendering/package.json
@@ -79,7 +79,7 @@
     "@types/compose-function": "^0.0.30",
     "@types/compression": "^0.0.36",
     "@types/crypto-js": "^4.1.1",
-    "@types/dompurify": "^2.0.4",
+    "@types/dompurify": "^2.3.3",
     "@types/express": "^4.17.11",
     "@types/gapi.analytics": "^0.0.3",
     "@types/google.analytics": "^0.0.41",
