+++ dotcom-rendering/src/model/sanitise.ts
@@ -1,9 +1,18 @@
+import type { Config } from 'dompurify';
 import createDOMPurify from 'dompurify';
 import { JSDOM } from 'jsdom';
 
 const { window } = new JSDOM('');
+// @ts-ignore -- Close enough types: DOMWindow is missing "self" from Window
 const DOMPurify = createDOMPurify(window);
 
-export const sanitiseHTML = (html: string, opts?: any): string =>
-	// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+type Opts = Config & {
+	RETURN_DOM_FRAGMENT?: false | undefined;
+	RETURN_DOM?: false | undefined;
+};
+
+export const sanitiseHTML = (html: string, opts: Opts = {}): string =>
 	DOMPurify.sanitize(html, opts);
+
+export const stripHTML = (s: string): string =>
+	DOMPurify.sanitize(s, { ALLOWED_TAGS: [] });
